HookerZ
Jonas will need to see deeper to overcome his obstacles and achieve his goals. As stated in the text, “The Receiver must have one more quality… The Capacity to See Beyond (Lowry, 63).” Essentially, The Capacity to See Beyond likely refers to seeing deeper and understanding the complexity and interconnectedness of the Community and the outside world, if he gets to see it. He already demonstrates different perspectives when he sees the apple and the audience change. He will need to see beyond the Community, and the system he is used to, to see the connections so that he can make better decisions for his Community. In the text, Jonas says, “‘I think it’s true,’ he told the Chief Elder and the community. ‘I don’t understand it yet. I don’t know what it is. But sometimes I see something. And maybe it’s beyond’ (Lowry, 64).” To illustrate, Jonas is referring to just seeing the audience change. He will need training to understand and use his ability. He will need to refine his ability to see deeper to become a good leader and manager for the community. It will be vital for his training, and for him to become a Receiver, one of his main obstacles. To summarize, Jonas will need to see deeper to understand the connectedness and dependencies of the community, and even the area outside the community, so he can make better decisions as the Receiver of Memory, a main obstacle, but he will need training.